+++ api/database.json
@@ -148,7 +148,11 @@
       "id": 3,
       "userId": 2,
       "message": "Help Me!?"
+    },
+    {
+      "userId": 1,
+      "message": "hi",
+      "id": 4
     }
- 
   ]
 }
