+++ src/components/messages/list.css
@@ -1,6 +1,17 @@
+.messages {
+    border: 1pt black solid;
+    max-width: 40%;
+    justify-content: center;
+   
+    
+}
 .message {
     border: 1pt solid;
     background-color: antiquewhite;
     padding: .5em;
     margin: 1.5em;
+}
+.messageForm {
+    display: flex;
+    align-items: baseline;
 }
